+++ packages/cloud_firestore_odm/cloud_firestore_odm_generator/lib/src/collection_data.dart
@@ -114,8 +114,9 @@ class CollectionData with Names {
 
     // TODO(validate name)
 
-    final path = annotation.getField('path')!.toStringValue()!;
+    final path = annotation.getField('path')!.toStringValue();
     _assertIsValidCollectionPath(path, annotatedElement);
+    path!;
 
     final type = CollectionData.modelTypeOfAnnotation(annotation);
 
@@ -160,7 +161,7 @@ class CollectionData with Names {
       );
     }
 
-    final annotatedElementSource = annotatedElement.librarySource!;
+    final annotatedElementSource = annotatedElement.librarySource;
 
     // TODO(rrousselGit) handle parts
     // Whether the model class and the reference variable are defined in the same file
@@ -305,8 +306,8 @@ represents the content of the collection must be in the same file.
             ?.contains(classPrefix) ??
         false) {
       throw InvalidGenerationSourceError(
-        'Defined a collection with duplicate class prefix $classPrefix.'
-        ' Either use a different class, or set a unique class prefix.',
+        'Defined a collection with duplicate class prefix $classPrefix. '
+        'Either use a different class, or set a unique class prefix.',
       );
     }
 
@@ -317,12 +318,20 @@ represents the content of the collection must be in the same file.
     return data;
   }
 
-  static void _assertIsValidCollectionPath(String path, Element element) {
+  static void _assertIsValidCollectionPath(String? path, Element element) {
+    if (path == null) {
+      throw InvalidGenerationSourceError(
+        'The annotation @Collection received "$path" as collection path, '
+        'but the path was null, which is not allowed.',
+        element: element,
+      );
+    }
+
     final allowedCharactersRegex = RegExp(r'^[0-9a-zA-Z/*-_]+$');
     if (!allowedCharactersRegex.hasMatch(path)) {
       throw InvalidGenerationSourceError(
         'The annotation @Collection received "$path" as collection path, '
-        ' but the path contains illegal characters.',
+        'but the path contains illegal characters.',
         element: element,
       );
     }
@@ -331,7 +340,7 @@ represents the content of the collection must be in the same file.
     if (pathSplit.length.isEven) {
       throw InvalidGenerationSourceError(
         'The annotation @Collection received "$path" as collection path, '
-        ' but this path points to a document instead of a collection',
+        'but this path points to a document instead of a collection',
         element: element,
       );
     }
@@ -340,7 +349,7 @@ represents the content of the collection must be in the same file.
       if (pathSplit[i] == '*') {
         throw InvalidGenerationSourceError(
           'The annotation @Collection received "$path" as collection path, '
-          ' but ${pathSplit[i]} is not a valid collection name',
+          'but ${pathSplit[i]} is not a valid collection name',
           element: element,
         );
       }