+++ packages/firebase_app_check/firebase_app_check_platform_interface/test/method_channel_tests/method_channel_firebase_app_check_test.dart
@@ -94,7 +94,7 @@ void main() {
             'FirebaseAppCheck#getToken',
             arguments: {
               'appName': defaultFirebaseAppName,
-              'forceRefresh': true
+              'forceRefresh': true,
             },
           ),
         ],