+++ packages/firebase_auth/firebase_auth_platform_interface/pigeons/messages.dart
@@ -17,8 +17,9 @@ import 'package:pigeon/pigeon.dart';
       package: 'io.flutter.plugins.firebase.auth',
       className: 'GeneratedAndroidFirebaseAuth',
     ),
-    objcHeaderOut: '../firebase_auth/ios/Classes/Public/messages.g.h',
-    objcSourceOut: '../firebase_auth/ios/Classes/messages.g.m',
+    objcHeaderOut:
+        '../firebase_auth/ios/Classes/Public/firebase_auth_messages.g.h',
+    objcSourceOut: '../firebase_auth/ios/Classes/firebase_auth_messages.g.m',
   ),
 )
 class PigeonMultiFactorSession {
@@ -128,11 +129,13 @@ class PigeonAdditionalUserInfo {
     required this.providerId,
     required this.username,
     this.profile,
+    this.authorizationCode,
   });
 
   final bool isNewUser;
   final String? providerId;
   final String? username;
+  final String? authorizationCode;
   final Map<String?, Object?>? profile;
 }
 
@@ -399,6 +402,11 @@ abstract class FirebaseAuthHostApi {
   String verifyPhoneNumber(
     PigeonFirebaseApp app,
     PigeonVerifyPhoneNumberRequest request,
+  );
+  @async
+  void revokeTokenWithAuthorizationCode(
+    PigeonFirebaseApp app,
+    String authorizationCode,
   );
 }